% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_table.R
\name{dp_table}
\alias{dp_table}
\title{function for tabulation}
\usage{
dp_table(dp, link = NULL, width = "100\%", height = "400px")
}
\arguments{
\item{dp}{an object of type diffrproject}

\item{link}{which link to produce table for}

\item{width}{width of widget}

\item{height}{heigth of widget}
}
\description{
function for tabulation
}

