% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_table.R
\name{dp_tableOutput}
\alias{dp_tableOutput}
\title{dp_table shiny output function}
\usage{
dp_tableOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{I have no idea}

\item{width}{width}

\item{height}{height}
}
\description{
dp_table shiny output function
}

