% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_vis.R
\name{dp_visOutput}
\alias{dp_visOutput}
\title{dp_vis shiny output function}
\usage{
dp_visOutput(outputId, width = "100\%", height = "400px")
}
\arguments{
\item{outputId}{I have no idea}

\item{width}{width}

\item{height}{height}
}
\description{
dp_vis shiny output function
}

