% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_table.R
\name{renderDp_table}
\alias{renderDp_table}
\title{dp_table shiny render function}
\usage{
renderDp_table(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{expr}

\item{env}{env}

\item{quoted}{quoted}
}
\description{
dp_table shiny render function
}

