% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_vis.R
\name{renderDP_vis}
\alias{renderDP_vis}
\title{dp_vis shiny render function}
\usage{
renderDP_vis(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{expr}{expr}

\item{env}{env}

\item{quoted}{quoted}
}
\description{
dp_vis shiny render function
}

