% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dp_vis.R
\name{dp_vis}
\alias{dp_vis}
\title{function for visualization}
\usage{
dp_vis(dp, width = "100\%", height = "400px")
}
\arguments{
\item{dp}{diffrprojects object}

\item{width}{width of widget}

\item{height}{heigth of widget}
}
\description{
function for visualization
}

